+++ src/util/condense_brackets/mod.rs
@@ -1,0 +1,5 @@
+mod condense;
+mod walk_ending;
+
+#[cfg(test)]
+mod tests;
