+++ src/util/condense_brackets/tests/walk_ending/mod.rs
@@ -1,0 +1,5 @@
+mod brace_multi;
+mod brace_strings;
+mod contain_brace;
+mod no_brace;
+mod same_line;
